refactor(config): remove raw events reprocessing configuration and service; update raw event consumption settings
