+++ internal/config/config.go
@@ -52,7 +52,6 @@ type Configuration struct {
 	WalletBalanceAlert         WalletBalanceAlertConfig         `mapstructure:"wallet_balance_alert" validate:"required"`
 	CustomerPortal             CustomerPortalConfig             `mapstructure:"customer_portal" validate:"required"`
 	Redis                      RedisConfig                      `mapstructure:"redis" validate:"required"`
-	RawEventsReprocessing      RawEventsReprocessingConfig      `mapstructure:"raw_events_reprocessing" validate:"required"`
 	RawEventConsumption        RawEventConsumptionConfig        `mapstructure:"raw_event_consumption" validate:"required"`
 }
 
@@ -272,11 +271,6 @@ type WalletBalanceAlertConfig struct {
 	Topic         string `mapstructure:"topic" default:"wallet_alert"`
 	RateLimit     int64  `mapstructure:"rate_limit" default:"1"`
 	ConsumerGroup string `mapstructure:"consumer_group" default:"v1_wallet_alert_service"`
-}
-
-type RawEventsReprocessingConfig struct {
-	Enabled     bool   `mapstructure:"enabled" default:"true"`
-	OutputTopic string `mapstructure:"output_topic" default:"prod_events_v4"`
 }
 
 type RawEventConsumptionConfig struct {
