+++ internal/config/config.yaml
@@ -270,15 +270,10 @@ redis:
   timeout: 5s # Timeout for Redis operations
   key_prefix: "flexprice:local"
 
-# Raw events reprocessing configuration
-raw_events_reprocessing:
-  enabled: true
-  output_topic: "prod_events_v4"
-
-# Raw event consumption configuration (batch processing from Bento)
+# Raw event consumption configuration (batch processing from Bento + reprocessing)
 raw_event_consumption:
-  enabled: true
+  enabled: false
   topic: "staging_events_backfill"      # INPUT: Reads batches from here
-  output_topic: "staging_events"                 # OUTPUT: Publishes transformed events to here
+  output_topic: "staging_events"         # OUTPUT: Publishes transformed events to here
   rate_limit: 10
   consumer_group: "v1_raw_event_processing_v4"
